+++ src/main/java/com/moilioncircle/redis/replicator/cmd/impl/IncrByCommand.java
@@ -7,12 +7,12 @@ import com.moilioncircle.redis.replicator.cmd.Command;
  */
 public class IncrByCommand implements Command {
     private String key;
-    private int value;
+    private long value;
 
     public IncrByCommand() {
     }
 
-    public IncrByCommand(String key, int value) {
+    public IncrByCommand(String key, long value) {
         this.key = key;
         this.value = value;
     }
@@ -25,11 +25,11 @@ public class IncrByCommand implements Command {
         this.key = key;
     }
 
-    public int getValue() {
+    public long getValue() {
         return value;
     }
 
-    public void setValue(int value) {
+    public void setValue(long value) {
         this.value = value;
     }
